* added latest version to all demo projects

+++ cocoa/HOWTO/SimpleSubscribe/PubNubMacOSTestApplication/libs/PubNub/Network/PNReachability.m
@@ -126,7 +126,7 @@ PNReachabilityStatus PNReachabilityStatusForFlags(SCNetworkReachabilityFlags fla
 
                 status = PNReachabilityStatusReachableViaWiFi;
 
-                unsigned long flagsForCleanUp = (unsigned long)flags;
+                unsigned int flagsForCleanUp = (unsigned int)flags;
                 PNBitsOff(&flagsForCleanUp, kSCNetworkReachabilityFlagsReachable, kSCNetworkReachabilityFlagsIsDirect,
                                             kSCNetworkReachabilityFlagsIsLocalAddress, BITS_LIST_TERMINATOR);
                 flags = (SCNetworkReachabilityFlags)flagsForCleanUp;
